Launch screen find bike (#1)
Initial implementation of main app screens for choosing what you want to search for, what location to use with search, and a card view with some details. Also implemented basic navigation through the main flow for searching for a bike using the device's location using mock data and simulated retrieval of location.